refactor tests into reasonable locations
